+++ src/main/java/com/codeQuartette/myTime/repository/impl/MyDateBulkRepository.java
@@ -1,4 +1,4 @@
-package com.codeQuartette.myTime.repository;
+package com.codeQuartette.myTime.repository.impl;
 
 import com.codeQuartette.myTime.domain.MyDate;
 import lombok.RequiredArgsConstructor;
